requirements.txt
PyPDF2==3.0.1
langchain
langchain_google_genai
google-generativeai
langchain_community
python-dotenv
streamlit
faiss-cpu
